reorganize a bit of early library and init code
Remove base.jl, since it no longer had a clear purpose. Its contents
are moved to various more specific files.

Move docs for Core from helpdb to basedocs.jl.
