+++ base/error.jl
@@ -85,7 +85,7 @@ systemerror(p, b::Bool; extrainfo=nothing) = b ? throw(Main.Base.SystemError(str
 Throw an [`AssertionError`](@ref) if `cond` is `false`.
 Also available as the macro [`@assert`](@ref).
 """
-assert(x) = x ? nothing : throw(Main.Base.AssertionError())
+assert(x) = x ? nothing : throw(AssertionError())
 
 """
     @assert cond [text]
@@ -114,7 +114,7 @@ macro assert(ex, msgs...)
         # string() might not be defined during bootstrap
         msg = :(Main.Base.string($(Expr(:quote,msg))))
     end
-    return :($(esc(ex)) ? $(nothing) : throw(Main.Base.AssertionError($msg)))
+    return :($(esc(ex)) ? $(nothing) : throw(AssertionError($msg)))
 end
 
 struct ExponentialBackOff
